Break up adamantium.rb in smaller chunks of code
