+++ config/flay.yml
@@ -1,3 +1,3 @@
 ---
 threshold: 9
-total_score: 29
+total_score: 32
