+++ config/flog.yml
@@ -1,2 +1,2 @@
 ---
-threshold: 23.4
+threshold: 20.3
